Cleanup GQL Cache update function files (#10)
