+++ src/graphql/cache/index.ts
@@ -1,1 +1,4 @@
-export * from './updateFunctions';
+export * from './AccountCache';
+export * from './BillsCache';
+export * from './NotesCache';
+export * from './PaymentsDueCache';
