+++ src/scss/components/_forecast_table.scss
@@ -28,8 +28,6 @@
       left: 0;
       position: sticky;
       z-index: 2;
-      background-color: $background;
-      background-image: $background-image;
     }
 
     .MuiTablePagination-spacer {
