+++ src/utils/selectors/index.ts
@@ -1,2 +1,2 @@
-export * from './GQLCache';
 export * from './account';
+export * from './GQLErrors';
